# Projektskibidi/apps/server/openapi.yaml
openapi: 3.1.0
info:
  title: AniCrawl Headless CMS API
  description: |
    A complete Headless CMS API for managing anime content.
    
    ## Authentication
    
    The API supports two authentication methods:
    
    ### API Key (Recommended for programmatic access)
    Include your API key in the `X-API-Key` header:
    ```
    X-API-Key: ak_your_api_key_here
    ```
    
    ### Session Cookie (For browser-based access)
    Login via `/api/auth/login` to receive a session cookie.
    
    ## Rate Limiting
    
    API keys have configurable rate limits (default: 1000 requests/hour).
    When exceeded, the API returns HTTP 429.
    
    ## Webhooks
    
    Subscribe to content events by creating webhooks. Supported events:
    - `anime.created`, `anime.updated`, `anime.deleted`
    - `user.created`, `user.updated`, `user.deleted`
    
  version: 1.0.0
  contact:
    name: AniCrawl Team
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    CookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: Session cookie from login
...
security:
  - ApiKeyAuth: []
  - CookieAuth: []
paths:
...
  /cms/media:
    get:
      tags: [Media]
      summary: List media files
      parameters:
        - name: folder
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [image, video, audio, document]
        - name: q
          in: query
          schema:
            type: string
      responses:
        200:
          description: List of media files
    post:
      tags: [Media]
      summary: Upload file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                alt:
                  type: string
                caption:
                  type: string
                folder:
                  type: string
      responses:
        201:
          description: File uploaded
  /cms/media/{id}:
    get:
      tags: [Media]
      summary: Get media file info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Media details
    patch:
      tags: [Media]
      summary: Update media metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Media updated
    delete:
      tags: [Media]
      summary: Delete media file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: Deleted
...